Fixed 64 bit math errors (ref. e9a6df7f44bc7ff798ee7d08d28e4563a4a54d1a). Changed format specifier to account for Android quirks.

+++ saturatr-android/jni/socket.hh
@@ -38,7 +38,7 @@ public:
     {}
 
     Packet( const Address &s_addr, const std::string &s_payload, const struct timeval &tv )
-      : addr( s_addr ), payload( s_payload ), timestamp( tv.tv_sec * 1000000000 + tv.tv_usec*1000 )
+      : addr( s_addr ), payload( s_payload ), timestamp( tv.tv_sec * ((uint64_t)1000000000) + tv.tv_usec* ((uint64_t)1000) )
     {}
   };
